Set up Project and Initial Test
Create and establish server. Createindex.spec.js to house tests.
Implement Test Endpoint Test which checks to see if a created
test endpoint returns a given message, in the right format, and
with the correct status code.

create Test Endpoint which sends DO NOT TEST ME BRO! to th
client.